TEMP: use 12.0.1, update to 13.x in finalize
Change recommended for standardization purposes (currently soundhax and ntrboot use 12.0.1 and don't prompt people to update which is inconsistent)
At the moment, 13.x is unable to boot on at least 10.6 and below.  So, to ensure that guide pages are (at least mostly) consistent, using 12.0.1 on all* pages and then prompting user to update luma in finalizing setup. This involves reorganizing finalizing setup a bit so that the cfw check and system update come before we modify the sd card.

Note: Change currently not copied to kartdlphax as this would involve reverting to an older commit and would cause complications with 3gx 12.0.1.